+++ apps/client/src/app/pipes/pipes/node-type-icon.pipe.ts
@@ -9,8 +9,8 @@ export class NodeTypeIconPipe implements PipeTransform {
     './assets/icons/nodes/060437.png',
     './assets/icons/nodes/060433.png',
     './assets/icons/nodes/060432.png',
-    './assets/icons/nodes/060465.png',
-    './assets/icons/nodes/060445.png'
+    './assets/icons/nodes/060445.png',
+    './assets/icons/nodes/060465.png'
   ];
 
   public static timed_icons = [
@@ -18,8 +18,8 @@ export class NodeTypeIconPipe implements PipeTransform {
     './assets/icons/nodes/060463.png',
     './assets/icons/nodes/060462.png',
     './assets/icons/nodes/060461.png',
-    './assets/icons/nodes/060466.png',
-    './assets/icons/nodes/060445.png'
+    './assets/icons/nodes/060445.png',
+    './assets/icons/nodes/060466.png'
   ];
 
   transform(type: number, timed?: boolean): string {