+++ apps/client/src/app/pipes/pipes/tug-name.pipe.ts
@@ -6,7 +6,10 @@ import { Tug } from '../../core/data/model/tug';
 })
 export class TugNamePipe implements PipeTransform {
 
-  transform(tug: Tug): string {
+  transform(tug: Tug): string | null {
+    if (!tug) {
+      return null;
+    }
     return `DB.FISH.TUG.${['Medium', 'Big', 'Light'][tug]}`;
   }