+++ apps/client/src/environments/patch-notes.ts
@@ -1,3 +1,21 @@
 export const patchNotes = `### Bug Fixes
 
-* **desktop:** fixed critical crash when completing a list item ingame.`;
+* **db:** fixed fishng icon on some items.
+* **gearsets:** fixed specialist flag always set to true when using in simulator.
+* **list:** fixed display not updating after filling from inventory.
+* **list:** fixed HQ requirement flag sometimes not saved properly.
+* **list:** fixes community checkbox not applying the change properly.
+* **log:** fixed Rime Dolomite now shown in the right category.
+* **search:** applying filters now applies query properly with debounce off.
+* **simulator:** fixed good actions rotation tip showing even when using Heart and Soul.
+* **ui:** fixed tug name sometimes shown as undefined.
+
+
+### Features
+
+* **db:** new hooksets data matrix on fishing spots.
+
+
+### Performance Improvements
+
+* **db:** removed fish ranking on desktop app to avoid loading freezes.`;